app/shared/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "podcast-automator-shared"
description = "Shared utilities for podcast automation system"
version = "0.1.0"
authors = [
    {name = "Your Name", email = "your-email@example.com"}
]
requires-python = ">=3.11"

dependencies = [
    "google-cloud-storage>=2.10",
    "google-cloud-aiplatform>=1.25",
    "google-cloud-secret-manager>=2.16",
    "google-cloud-logging>=3.5",
    "boto3>=1.26",
    "requests>=2.31",
]

[tool.setuptools]
packages = ["."]

[tool.black]
line-length = 100
target-version = ['py311']
